Move the player and map owners' receiver-free rules out (#678)
Fifth semantic vertical, applying #676's pattern to the other curated
owners. Twelve receiver-free functions left two owners:

- `wow-entities/src/player` loses 3 functions; its aggregate drops 11,374 to
  11,339 production lines (21,564 to 21,529 total).
- `wow-map/src/map` loses 9; its aggregate drops 16,168 to 16,055 production
  lines (34,896 to 34,560 total).

Both ledger baselines are tightened to the validated figures. 725
wow-entities tests and 727 wow-map tests pass unchanged.

## What the survey actually found, and what it cost

The first pass counted 111 candidates across six owners. Most were
constructors: a receiver-free `fn new() -> Self` or `default()` is not a
rule, it is the type's own vocabulary, and moving one produced 480 errors
from names colliding across types. The detector now excludes anything whose
signature mentions `Self` and any name defined on more than one type, which
leaves 31 - and of those, only these twelve could move.

The rest did not, for reasons worth recording:

- The one `wow-social/src/group` candidate reads `PendingInvites`' private
  field. It is group-internal behaviour that happens to take no receiver, so
  moving it would have meant widening that field. Reverted.
- The 18 candidates in `handlers/{quest,character,loot}` read the quest
  owner's private vocabulary - 17 module-level `*_LIKE_CPP_LOCAL` constants
  and `super::`-relative paths. Moving the functions without that vocabulary
  turned constant match arms into binding patterns, which the compiler
  caught as E0408 rather than silently. They need a delivery scoped to that
  owner, which can move the constants with them.

One rule crossed a crate boundary: `wow-world` calls
`Player::is_using_two_handed_weapon_in_one_hand_template`, so that function
is `pub` in a `pub mod player_rules` and the call site imports it under an
alias, keeping the consumer's path length unchanged.

Architecture check, self-test, physical ratchet, hotspot ratchet and the
syntax-only session ownership check pass; the workspace build is clean.

Closes #678
